# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScenePulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: listWatchlist
      tags: [watchlist]
      summary: List the current user's watchlisted venues
      responses:
        "200":
          description: Watchlisted venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
        "401":
          description: Not authenticated
  /watchlist/{venueId}:
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add a venue to the current user's watchlist
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Added (idempotent — also returned if already present)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistEntry"
        "401":
          description: Not authenticated
        "404":
          description: Venue not found
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove a venue from the current user's watchlist
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
        "401":
          description: Not authenticated
  /me/activity:
    get:
      operationId: getMyActivity
      tags: [watchlist]
      summary: Current user's past reports and comments
      responses:
        "200":
          description: The user's activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyActivity"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Venue:
      type: object
      required:
        - id
        - name
        - category
        - city
        - market
        - region
        - country
        - latitude
        - longitude
        - address
        - rating
        - crowdScore
        - crowdLevel
        - waitTimeMinutes
        - headcount
        - lineTrend
        - seatingOdds
        - noiseLevel
        - coverCost
        - bestTimeWindow
        - bestFor
        - operatorGapNote
        - peakPressureWindow
        - reservationSignal
        - staffingSignal
        - dataSignalsTracked
        - arrivalTips
        - isWatchlisted
        - updatedAt
      properties:
        id: { type: integer }
        name: { type: string }
        category:
          type: string
          enum: [bar, restaurant, retail, cafe, experience]
        city: { type: string }
        market: { type: string }
        region: { type: string }
        country: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        address: { type: string }
        rating: { type: number }
        crowdScore:
          type: integer
          description: 0-100 crowd pressure score
        crowdLevel:
          type: string
          enum: [open, lively, packed]
        waitTimeMinutes: { type: integer }
        headcount: { type: integer }
        lineTrend:
          type: string
          enum: [rising, falling, steady]
        seatingOdds: { type: string }
        noiseLevel:
          type: string
          enum: [quiet, moderate, loud]
        coverCost: { type: string }
        bestTimeWindow: { type: string }
        bestFor:
          type: array
          items: { type: string }
        operatorGapNote: { type: string }
        peakPressureWindow: { type: string }
        reservationSignal: { type: string }
        staffingSignal: { type: string }
        dataSignalsTracked:
          type: array
          items: { type: string }
        arrivalTips:
          type: array
          items: { type: string }
        sourceLabel: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        mapsUrl: { type: string }
        isWatchlisted: { type: boolean }
        updatedAt: { type: string, format: date-time }
...
    Comment:
      type: object
      required: [id, venueId, authorName, message, likes, dislikes, createdAt]
      properties:
        id: { type: integer }
        venueId: { type: integer }
        parentCommentId: { type: [integer, "null"] }
        authorName: { type: string }
        message: { type: string }
        likes: { type: integer }
        dislikes: { type: integer }
        createdAt: { type: string, format: date-time }
...
    LiveReport:
      type: object
      required:
        - id
        - venueId
        - reporterName
        - crowdLevel
        - waitTimeMinutes
        - vibeNote
        - createdAt
      properties:
        id: { type: integer }
        venueId: { type: integer }
        reporterName: { type: string }
        crowdLevel:
          type: string
          enum: [open, lively, packed]
        waitTimeMinutes: { type: integer }
        vibeNote: { type: string }
        createdAt: { type: string, format: date-time }
    WatchlistEntry:
      type: object
      required: [id, userId, venueId, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        venueId: { type: integer }
        createdAt: { type: string, format: date-time }
    MyActivity:
      type: object
      required: [reports, comments]
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/MyReport"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/MyComment"
    MyReport:
      allOf:
        - $ref: "#/components/schemas/LiveReport"
        - type: object
          required: [venueName]
          properties:
            venueName: { type: string }
    MyComment:
      allOf:
        - $ref: "#/components/schemas/Comment"
        - type: object
          required: [venueName]
          properties:
            venueName: { type: string }
...