# lib/api-zod/src/generated/types/myActivity.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ScenePulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MyComment } from './myComment';
import type { MyReport } from './myReport';

export interface MyActivity {
  reports: MyReport[];
  comments: MyComment[];
}