# lib/api-zod/src/generated/types/myComment.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ScenePulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from './comment';

export type MyComment = Comment & {
  venueName: string;
};