# lib/api-zod/src/generated/types/myReport.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ScenePulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LiveReport } from './liveReport';

export type MyReport = LiveReport & {
  venueName: string;
};