# lib/api-zod/src/generated/types/watchlistEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ScenePulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistEntry {
  id: number;
  userId: string;
  venueId: number;
  createdAt: Date;
}